moved branch popup out to own module
