+++ app/src/ui/toolbar/branch-dropdown.tsx
@@ -5,7 +5,7 @@ import { Repository } from '../../models/repository'
 import { TipState } from '../../models/tip'
 import { ToolbarDropdown, DropdownState } from './dropdown'
 import { IRepositoryState } from '../../lib/app-state'
-import { Branches } from '../branches'
+import { Branches } from '../branches/branch-popup'
 import { assertNever } from '../../lib/fatal-error'
 import { BranchesTab } from '../../models/branches-tab'
 import { enablePRIntegration } from '../../lib/feature-flag'
